build with snapshot

+++ src/test/kotlin/commands/ngrok-server.kts
@@ -73,5 +73,5 @@ runBlocking {
     }
   }
 
-  delay(Int.MAX_VALUE)
+  delay(Long.MAX_VALUE)
 }
